chore: comments, renames, and cleanup
